requirements.txt
psycopg2-binary>=2.9
rich>=12.0; python_version >= '3.6'
python-dotenv>=0.21